config ecommerce service entities, create service api

+++ src/EcommeceService/Domain/Aggregates/Services/Group.cs
@@ -4,10 +4,9 @@ namespace Domain.Aggregates.Services
 {
     public class Group : BaseEntity
     {
-
         public string Name { get; set; } = default!;
         public string? Description { get; set; }
-        public long Price { get; set; } = default!;
+        public decimal Price { get; set; } = default!;
         public bool Disable { get; set; } = default!;
 
         public IEnumerable<GroupService> GroupServices { get; set; } = [];
